LIGHT test.
